Move notifications to package
