+++ 3ds/add_friendship_by_pid.go
@@ -3,6 +3,7 @@ package friends_3ds
 import (
 	database_3ds "github.com/PretendoNetwork/friends-secure/database/3ds"
 	"github.com/PretendoNetwork/friends-secure/globals"
+	notifications_3ds "github.com/PretendoNetwork/friends-secure/notifications/3ds"
 	nex "github.com/PretendoNetwork/nex-go"
 	nexproto "github.com/PretendoNetwork/nex-protocols-go"
 )
@@ -12,7 +13,7 @@ func AddFriendshipByPrincipalID(err error, client *nex.Client, callID uint32, lf
 
 	connectedUser := globals.ConnectedUsers[pid]
 	if connectedUser != nil {
-		go sendFriendshipCompletedNotification(connectedUser.Client, pid, client.PID())
+		go notifications_3ds.SendFriendshipCompleted(connectedUser.Client, pid, client.PID())
 	}
 
 	rmcResponseStream := nex.NewStreamOut(globals.NEXServer)
@@ -38,42 +39,4 @@ func AddFriendshipByPrincipalID(err error, client *nex.Client, callID uint32, lf
 	responsePacket.AddFlag(nex.FlagReliable)
 
 	globals.NEXServer.Send(responsePacket)
-}
-
-func sendFriendshipCompletedNotification(client *nex.Client, friendPID uint32, senderPID uint32) {
-	notificationEvent := nexproto.NewNintendoNotificationEventGeneral()
-	notificationEvent.U32Param = 0
-	notificationEvent.U64Param1 = 0
-	notificationEvent.U64Param2 = uint64(friendPID)
-
-	eventObject := nexproto.NewNintendoNotificationEvent()
-	eventObject.Type = 7
-	eventObject.SenderPID = senderPID
-	eventObject.DataHolder = nex.NewDataHolder()
-	eventObject.DataHolder.SetTypeName("NintendoNotificationEventGeneral")
-	eventObject.DataHolder.SetObjectData(notificationEvent)
-
-	stream := nex.NewStreamOut(globals.NEXServer)
-	eventObjectBytes := eventObject.Bytes(stream)
-
-	rmcRequest := nex.NewRMCRequest()
-	rmcRequest.SetProtocolID(nexproto.NintendoNotificationsProtocolID)
-	rmcRequest.SetCallID(3810693103)
-	rmcRequest.SetMethodID(nexproto.NintendoNotificationsMethodProcessNintendoNotificationEvent1)
-	rmcRequest.SetParameters(eventObjectBytes)
-
-	rmcRequestBytes := rmcRequest.Bytes()
-
-	requestPacket, _ := nex.NewPacketV0(client, nil)
-
-	requestPacket.SetVersion(0)
-	requestPacket.SetSource(0xA1)
-	requestPacket.SetDestination(0xAF)
-	requestPacket.SetType(nex.DataPacket)
-	requestPacket.SetPayload(rmcRequestBytes)
-
-	requestPacket.AddFlag(nex.FlagNeedsAck)
-	requestPacket.AddFlag(nex.FlagReliable)
-
-	globals.NEXServer.Send(requestPacket)
 }
