+++ 3ds/update_mii.go
@@ -3,12 +3,13 @@ package friends_3ds
 import (
 	database_3ds "github.com/PretendoNetwork/friends-secure/database/3ds"
 	"github.com/PretendoNetwork/friends-secure/globals"
+	notifications_3ds "github.com/PretendoNetwork/friends-secure/notifications/3ds"
 	nex "github.com/PretendoNetwork/nex-go"
 	nexproto "github.com/PretendoNetwork/nex-protocols-go"
 )
 
 func UpdateMii(err error, client *nex.Client, callID uint32, mii *nexproto.Mii) {
-	go sendMiiUpdateNotification(client)
+	go notifications_3ds.SendMiiUpdateNotification(client)
 	database_3ds.UpdateUserMii(client.PID(), mii)
 
 	rmcResponse := nex.NewRMCResponse(nexproto.Friends3DSProtocolID, callID)
@@ -28,48 +29,4 @@ func UpdateMii(err error, client *nex.Client, callID uint32, mii *nexproto.Mii) 
 	responsePacket.AddFlag(nex.FlagReliable)
 
 	globals.NEXServer.Send(responsePacket)
-}
-
-func sendMiiUpdateNotification(client *nex.Client) {
-	notificationEvent := nexproto.NewNintendoNotificationEventGeneral()
-
-	eventObject := nexproto.NewNintendoNotificationEvent()
-	eventObject.Type = 5
-	eventObject.SenderPID = client.PID()
-	eventObject.DataHolder = nex.NewDataHolder()
-	eventObject.DataHolder.SetTypeName("NintendoNotificationEventGeneral")
-	eventObject.DataHolder.SetObjectData(notificationEvent)
-
-	stream := nex.NewStreamOut(globals.NEXServer)
-	eventObjectBytes := eventObject.Bytes(stream)
-
-	rmcRequest := nex.NewRMCRequest()
-	rmcRequest.SetProtocolID(nexproto.NintendoNotificationsProtocolID)
-	rmcRequest.SetCallID(3810693103)
-	rmcRequest.SetMethodID(nexproto.NintendoNotificationsMethodProcessNintendoNotificationEvent1)
-	rmcRequest.SetParameters(eventObjectBytes)
-
-	rmcRequestBytes := rmcRequest.Bytes()
-
-	friendsList := database_3ds.GetUserFriends(client.PID())
-
-	for i := 0; i < len(friendsList); i++ {
-
-		connectedUser := globals.ConnectedUsers[friendsList[i].PID]
-
-		if connectedUser != nil {
-			requestPacket, _ := nex.NewPacketV0(connectedUser.Client, nil)
-
-			requestPacket.SetVersion(0)
-			requestPacket.SetSource(0xA1)
-			requestPacket.SetDestination(0xAF)
-			requestPacket.SetType(nex.DataPacket)
-			requestPacket.SetPayload(rmcRequestBytes)
-
-			requestPacket.AddFlag(nex.FlagNeedsAck)
-			requestPacket.AddFlag(nex.FlagReliable)
-
-			globals.NEXServer.Send(requestPacket)
-		}
-	}
 }
