+++ 3ds/update_preference.go
@@ -3,6 +3,7 @@ package friends_3ds
 import (
 	database_3ds "github.com/PretendoNetwork/friends-secure/database/3ds"
 	"github.com/PretendoNetwork/friends-secure/globals"
+	notifications_3ds "github.com/PretendoNetwork/friends-secure/notifications/3ds"
 	nex "github.com/PretendoNetwork/nex-go"
 	nexproto "github.com/PretendoNetwork/nex-protocols-go"
 )
@@ -12,10 +13,10 @@ func UpdatePreference(err error, client *nex.Client, callID uint32, showOnline b
 		emptyPresence := nexproto.NewNintendoPresence()
 		emptyPresence.GameKey = nexproto.NewGameKey()
 		emptyPresence.ChangedFlags = 0xFFFFFFFF // All flags
-		sendPresenceUpdateNotification(client, emptyPresence)
+		notifications_3ds.SendPresenceUpdate(client, emptyPresence)
 	}
 	if !showOnline {
-		SendUserWentOfflineNotificationsGlobally(client)
+		notifications_3ds.SendUserWentOfflineGlobally(client)
 	}
 
 	database_3ds.UpdateUserPreferences(client.PID(), showOnline, showCurrentGame)
