+++ 3ds/update_presence.go
@@ -1,8 +1,8 @@
 package friends_3ds
 
 import (
-	database_3ds "github.com/PretendoNetwork/friends-secure/database/3ds"
 	"github.com/PretendoNetwork/friends-secure/globals"
+	notifications_3ds "github.com/PretendoNetwork/friends-secure/notifications/3ds"
 	nex "github.com/PretendoNetwork/nex-go"
 	nexproto "github.com/PretendoNetwork/nex-protocols-go"
 )
@@ -17,7 +17,7 @@ func UpdatePresence(err error, client *nex.Client, callID uint32, presence *nexp
 		currentPresence.ChangedFlags = 0xFFFFFFFF // All flags
 	}
 
-	go sendPresenceUpdateNotification(client, currentPresence)
+	go notifications_3ds.SendPresenceUpdate(client, currentPresence)
 	globals.ConnectedUsers[client.PID()].Presence = currentPresence
 
 	rmcResponse := nex.NewRMCResponse(nexproto.Friends3DSProtocolID, callID)
@@ -37,47 +37,4 @@ func UpdatePresence(err error, client *nex.Client, callID uint32, presence *nexp
 	responsePacket.AddFlag(nex.FlagReliable)
 
 	globals.NEXServer.Send(responsePacket)
-}
-
-func sendPresenceUpdateNotification(client *nex.Client, presence *nexproto.NintendoPresence) {
-	eventObject := nexproto.NewNintendoNotificationEvent()
-	eventObject.Type = 1
-	eventObject.SenderPID = client.PID()
-	eventObject.DataHolder = nex.NewDataHolder()
-	eventObject.DataHolder.SetTypeName("NintendoPresence")
-	eventObject.DataHolder.SetObjectData(presence)
-
-	stream := nex.NewStreamOut(globals.NEXServer)
-	eventObjectBytes := eventObject.Bytes(stream)
-
-	rmcRequest := nex.NewRMCRequest()
-	rmcRequest.SetProtocolID(nexproto.NintendoNotificationsProtocolID)
-	rmcRequest.SetCallID(3810693103)
-	rmcRequest.SetMethodID(nexproto.NintendoNotificationsMethodProcessNintendoNotificationEvent1)
-	rmcRequest.SetParameters(eventObjectBytes)
-
-	rmcRequestBytes := rmcRequest.Bytes()
-
-	friendsList := database_3ds.GetUserFriends(client.PID())
-
-	for i := 0; i < len(friendsList); i++ {
-
-		connectedUser := globals.ConnectedUsers[friendsList[i].PID]
-
-		if connectedUser != nil {
-
-			requestPacket, _ := nex.NewPacketV0(connectedUser.Client, nil)
-
-			requestPacket.SetVersion(0)
-			requestPacket.SetSource(0xA1)
-			requestPacket.SetDestination(0xAF)
-			requestPacket.SetType(nex.DataPacket)
-			requestPacket.SetPayload(rmcRequestBytes)
-
-			requestPacket.AddFlag(nex.FlagNeedsAck)
-			requestPacket.AddFlag(nex.FlagReliable)
-
-			globals.NEXServer.Send(requestPacket)
-		}
-	}
 }
