+++ notifications/3ds/send_favorite_update.go
@@ -1,0 +1,50 @@
+package notifications_3ds
+
+import (
+	database_3ds "github.com/PretendoNetwork/friends-secure/database/3ds"
+	"github.com/PretendoNetwork/friends-secure/globals"
+	nex "github.com/PretendoNetwork/nex-go"
+	nexproto "github.com/PretendoNetwork/nex-protocols-go"
+)
+
+func SendFavoriteUpdate(client *nex.Client, gameKey *nexproto.GameKey) {
+	eventObject := nexproto.NewNintendoNotificationEvent()
+	eventObject.Type = 2
+	eventObject.SenderPID = client.PID()
+	eventObject.DataHolder = nex.NewDataHolder()
+	eventObject.DataHolder.SetTypeName("GameKey")
+	eventObject.DataHolder.SetObjectData(gameKey)
+
+	stream := nex.NewStreamOut(globals.NEXServer)
+	eventObjectBytes := eventObject.Bytes(stream)
+
+	rmcRequest := nex.NewRMCRequest()
+	rmcRequest.SetProtocolID(nexproto.NintendoNotificationsProtocolID)
+	rmcRequest.SetCallID(3810693103)
+	rmcRequest.SetMethodID(nexproto.NintendoNotificationsMethodProcessNintendoNotificationEvent1)
+	rmcRequest.SetParameters(eventObjectBytes)
+
+	rmcRequestBytes := rmcRequest.Bytes()
+
+	friendsList := database_3ds.GetUserFriends(client.PID())
+
+	for i := 0; i < len(friendsList); i++ {
+
+		connectedUser := globals.ConnectedUsers[friendsList[i].PID]
+
+		if connectedUser != nil {
+			requestPacket, _ := nex.NewPacketV0(connectedUser.Client, nil)
+
+			requestPacket.SetVersion(0)
+			requestPacket.SetSource(0xA1)
+			requestPacket.SetDestination(0xAF)
+			requestPacket.SetType(nex.DataPacket)
+			requestPacket.SetPayload(rmcRequestBytes)
+
+			requestPacket.AddFlag(nex.FlagNeedsAck)
+			requestPacket.AddFlag(nex.FlagReliable)
+
+			globals.NEXServer.Send(requestPacket)
+		}
+	}
+}
