+++ notifications/3ds/send_user_went_offline.go
@@ -1,0 +1,55 @@
+package notifications_3ds
+
+import (
+	database_3ds "github.com/PretendoNetwork/friends-secure/database/3ds"
+	"github.com/PretendoNetwork/friends-secure/globals"
+	nex "github.com/PretendoNetwork/nex-go"
+	nexproto "github.com/PretendoNetwork/nex-protocols-go"
+)
+
+func SendUserWentOfflineGlobally(client *nex.Client) {
+	friendsList := database_3ds.GetUserFriends(client.PID())
+
+	for i := 0; i < len(friendsList); i++ {
+		SendUserWentOffline(client, friendsList[i].PID)
+	}
+}
+
+func SendUserWentOffline(client *nex.Client, pid uint32) {
+	notificationEvent := nexproto.NewNintendoNotificationEventGeneral()
+
+	eventObject := nexproto.NewNintendoNotificationEvent()
+	eventObject.Type = 10
+	eventObject.SenderPID = client.PID()
+	eventObject.DataHolder = nex.NewDataHolder()
+	eventObject.DataHolder.SetTypeName("NintendoNotificationEventGeneral")
+	eventObject.DataHolder.SetObjectData(notificationEvent)
+
+	stream := nex.NewStreamOut(globals.NEXServer)
+	eventObjectBytes := eventObject.Bytes(stream)
+
+	rmcRequest := nex.NewRMCRequest()
+	rmcRequest.SetProtocolID(nexproto.NintendoNotificationsProtocolID)
+	rmcRequest.SetCallID(3810693103)
+	rmcRequest.SetMethodID(nexproto.NintendoNotificationsMethodProcessNintendoNotificationEvent1)
+	rmcRequest.SetParameters(eventObjectBytes)
+
+	rmcRequestBytes := rmcRequest.Bytes()
+
+	connectedUser := globals.ConnectedUsers[pid]
+
+	if connectedUser != nil {
+		requestPacket, _ := nex.NewPacketV0(connectedUser.Client, nil)
+
+		requestPacket.SetVersion(0)
+		requestPacket.SetSource(0xA1)
+		requestPacket.SetDestination(0xAF)
+		requestPacket.SetType(nex.DataPacket)
+		requestPacket.SetPayload(rmcRequestBytes)
+
+		requestPacket.AddFlag(nex.FlagNeedsAck)
+		requestPacket.AddFlag(nex.FlagReliable)
+
+		globals.NEXServer.Send(requestPacket)
+	}
+}
