+++ notifications/wiiu/send_friend_request.go
@@ -1,0 +1,40 @@
+package notifications_wiiu
+
+import (
+	"github.com/PretendoNetwork/friends-secure/globals"
+	nex "github.com/PretendoNetwork/nex-go"
+	nexproto "github.com/PretendoNetwork/nex-protocols-go"
+)
+
+func SendFriendRequest(client *nex.Client, friendRequestNotificationData *nexproto.FriendRequest) {
+	eventObject := nexproto.NewNintendoNotificationEvent()
+	eventObject.Type = 27
+	eventObject.SenderPID = friendRequestNotificationData.PrincipalInfo.PID
+	eventObject.DataHolder = nex.NewDataHolder()
+	eventObject.DataHolder.SetTypeName("FriendRequest")
+	eventObject.DataHolder.SetObjectData(friendRequestNotificationData)
+
+	stream := nex.NewStreamOut(globals.NEXServer)
+	eventObjectBytes := eventObject.Bytes(stream)
+
+	rmcRequest := nex.NewRMCRequest()
+	rmcRequest.SetProtocolID(nexproto.NintendoNotificationsProtocolID)
+	rmcRequest.SetCallID(3810693103)
+	rmcRequest.SetMethodID(nexproto.NintendoNotificationsMethodProcessNintendoNotificationEvent2)
+	rmcRequest.SetParameters(eventObjectBytes)
+
+	rmcRequestBytes := rmcRequest.Bytes()
+
+	requestPacket, _ := nex.NewPacketV0(client, nil)
+
+	requestPacket.SetVersion(0)
+	requestPacket.SetSource(0xA1)
+	requestPacket.SetDestination(0xAF)
+	requestPacket.SetType(nex.DataPacket)
+	requestPacket.SetPayload(rmcRequestBytes)
+
+	requestPacket.AddFlag(nex.FlagNeedsAck)
+	requestPacket.AddFlag(nex.FlagReliable)
+
+	globals.NEXServer.Send(requestPacket)
+}
