+++ wiiu/accept_friend_request.go
@@ -3,6 +3,7 @@ package friends_wiiu
 import (
 	database_wiiu "github.com/PretendoNetwork/friends-secure/database/wiiu"
 	"github.com/PretendoNetwork/friends-secure/globals"
+	notifications_wiiu "github.com/PretendoNetwork/friends-secure/notifications/wiiu"
 	nex "github.com/PretendoNetwork/nex-go"
 	nexproto "github.com/PretendoNetwork/nex-protocols-go"
 )
@@ -26,7 +27,7 @@ func AcceptFriendRequest(err error, client *nex.Client, callID uint32, id uint64
 		senderFriendInfo.LastOnline = friendInfo.LastOnline // TODO: Change this
 		senderFriendInfo.Unknown = 0
 
-		go sendFriendRequestAcceptedNotification(connectedUser.Client, senderFriendInfo)
+		go notifications_wiiu.SendFriendRequestAccepted(connectedUser.Client, senderFriendInfo)
 	}
 
 	rmcResponseStream := nex.NewStreamOut(globals.NEXServer)
@@ -53,37 +54,4 @@ func AcceptFriendRequest(err error, client *nex.Client, callID uint32, id uint64
 	responsePacket.AddFlag(nex.FlagReliable)
 
 	globals.NEXServer.Send(responsePacket)
-}
-
-func sendFriendRequestAcceptedNotification(client *nex.Client, friendInfo *nexproto.FriendInfo) {
-	eventObject := nexproto.NewNintendoNotificationEvent()
-	eventObject.Type = 30
-	eventObject.SenderPID = friendInfo.NNAInfo.PrincipalBasicInfo.PID
-	eventObject.DataHolder = nex.NewDataHolder()
-	eventObject.DataHolder.SetTypeName("FriendInfo")
-	eventObject.DataHolder.SetObjectData(friendInfo)
-
-	stream := nex.NewStreamOut(globals.NEXServer)
-	eventObjectBytes := eventObject.Bytes(stream)
-
-	rmcRequest := nex.NewRMCRequest()
-	rmcRequest.SetProtocolID(nexproto.NintendoNotificationsProtocolID)
-	rmcRequest.SetCallID(3810693103)
-	rmcRequest.SetMethodID(nexproto.NintendoNotificationsMethodProcessNintendoNotificationEvent1)
-	rmcRequest.SetParameters(eventObjectBytes)
-
-	rmcRequestBytes := rmcRequest.Bytes()
-
-	requestPacket, _ := nex.NewPacketV0(client, nil)
-
-	requestPacket.SetVersion(0)
-	requestPacket.SetSource(0xA1)
-	requestPacket.SetDestination(0xAF)
-	requestPacket.SetType(nex.DataPacket)
-	requestPacket.SetPayload(rmcRequestBytes)
-
-	requestPacket.AddFlag(nex.FlagNeedsAck)
-	requestPacket.AddFlag(nex.FlagReliable)
-
-	globals.NEXServer.Send(requestPacket)
 }
