+++ wiiu/update_and_get_all_information.go
@@ -5,6 +5,7 @@ import (
 
 	database_wiiu "github.com/PretendoNetwork/friends-secure/database/wiiu"
 	"github.com/PretendoNetwork/friends-secure/globals"
+	notifications_wiiu "github.com/PretendoNetwork/friends-secure/notifications/wiiu"
 	nex "github.com/PretendoNetwork/nex-go"
 	nexproto "github.com/PretendoNetwork/nex-protocols-go"
 )
@@ -21,7 +22,7 @@ func UpdateAndGetAllInformation(err error, client *nex.Client, callID uint32, nn
 	presence.Online = true      // Force online status. I have no idea why this is always false
 	presence.PID = client.PID() // WHY IS THIS SET TO 0 BY DEFAULT??
 
-	sendUpdatePresenceWiiUNotifications(presence)
+	notifications_wiiu.SendPresenceUpdate(presence)
 
 	// Get user information
 	pid := client.PID()
