+++ wiiu/update_presence.go
@@ -1,10 +1,8 @@
 package friends_wiiu
 
 import (
-	"fmt"
-
-	database_wiiu "github.com/PretendoNetwork/friends-secure/database/wiiu"
 	"github.com/PretendoNetwork/friends-secure/globals"
+	notifications_wiiu "github.com/PretendoNetwork/friends-secure/notifications/wiiu"
 	nex "github.com/PretendoNetwork/nex-go"
 	nexproto "github.com/PretendoNetwork/nex-protocols-go"
 )
@@ -16,7 +14,7 @@ func UpdatePresence(err error, client *nex.Client, callID uint32, presence *nexp
 	presence.PID = client.PID() // WHY IS THIS SET TO 0 BY DEFAULT??
 
 	globals.ConnectedUsers[pid].PresenceV2 = presence
-	sendUpdatePresenceWiiUNotifications(presence)
+	notifications_wiiu.SendPresenceUpdate(presence)
 
 	rmcResponse := nex.NewRMCResponse(nexproto.FriendsWiiUProtocolID, callID)
 	rmcResponse.SetSuccess(nexproto.FriendsWiiUMethodUpdatePresence, nil)
@@ -35,69 +33,4 @@ func UpdatePresence(err error, client *nex.Client, callID uint32, presence *nexp
 	responsePacket.AddFlag(nex.FlagReliable)
 
 	globals.NEXServer.Send(responsePacket)
-}
-
-func sendUpdatePresenceWiiUNotifications(presence *nexproto.NintendoPresenceV2) {
-	eventObject := nexproto.NewNintendoNotificationEvent()
-	eventObject.Type = 24
-	eventObject.SenderPID = presence.PID
-	eventObject.DataHolder = nex.NewDataHolder()
-	eventObject.DataHolder.SetTypeName("NintendoPresenceV2")
-	eventObject.DataHolder.SetObjectData(presence)
-
-	stream := nex.NewStreamOut(globals.NEXServer)
-	eventObjectBytes := eventObject.Bytes(stream)
-
-	rmcRequest := nex.NewRMCRequest()
-	rmcRequest.SetProtocolID(nexproto.NintendoNotificationsProtocolID)
-	rmcRequest.SetCallID(3810693103)
-	rmcRequest.SetMethodID(nexproto.NintendoNotificationsMethodProcessNintendoNotificationEvent2)
-	rmcRequest.SetParameters(eventObjectBytes)
-
-	rmcRequestBytes := rmcRequest.Bytes()
-
-	friendList := database_wiiu.GetUserFriendList(presence.PID)
-
-	for i := 0; i < len(friendList); i++ {
-		if friendList[i] == nil || friendList[i].NNAInfo == nil || friendList[i].NNAInfo.PrincipalBasicInfo == nil {
-			// TODO: Fix this
-			pid := presence.PID
-			var friendPID uint32 = 0
-
-			if friendList[i] != nil && friendList[i].Presence != nil {
-				// TODO: Better track the bad users PID
-				friendPID = friendList[i].Presence.PID
-			}
-
-			globals.Logger.Error(fmt.Sprintf("User %d has friend %d with bad presence data", pid, friendPID))
-
-			if friendList[i] == nil {
-				globals.Logger.Error(fmt.Sprintf("%d friendList[i] nil", friendPID))
-			} else if friendList[i].NNAInfo == nil {
-				globals.Logger.Error(fmt.Sprintf("%d friendList[i].NNAInfo is nil", friendPID))
-			} else if friendList[i].NNAInfo.PrincipalBasicInfo == nil {
-				globals.Logger.Error(fmt.Sprintf("%d friendList[i].NNAInfo.PrincipalBasicInfo is nil", friendPID))
-			}
-
-			continue
-		}
-
-		friendPID := friendList[i].NNAInfo.PrincipalBasicInfo.PID
-		connectedUser := globals.ConnectedUsers[friendPID]
-
-		if connectedUser != nil {
-			requestPacket, _ := nex.NewPacketV0(connectedUser.Client, nil)
-
-			requestPacket.SetVersion(0)
-			requestPacket.SetSource(0xA1)
-			requestPacket.SetDestination(0xAF)
-			requestPacket.SetType(nex.DataPacket)
-			requestPacket.SetPayload(rmcRequestBytes)
-
-			requestPacket.AddFlag(nex.FlagNeedsAck)
-			requestPacket.AddFlag(nex.FlagReliable)
-
-			globals.NEXServer.Send(requestPacket)
-		}
-	}
 }
